medicare_backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
python-dotenv
gunicorn
whitenoise
drf-spectacular
dj-database-url
requests